requirements.txt
anthropic>=0.40.0
requests>=2.31.0
python-telegram-bot>=21.0
